# tf-api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Swagger HLF Trade Finance.",
    "version": "1.0.0",
    "title": "Swagger HLF Trade Finance",

    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "184.105.118.161:3000",
  "basePath": "/tf",

  "schemes": ["http"],
  "paths": {
    "/requestLC": {
      "post": {
        "summary": "Request LC",
        "description": "Request LC",
        "operationId": "requestLC",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LC"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/issueLC": {
      "post": {
        "summary": "Issue LC",
        "description": "Issue LC",
        "operationId": "issueLC",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC ID object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LCID"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/acceptLC": {
      "post": {
        "summary": "Accept LC",
        "description": "Accept LC",
        "operationId": "acceptLC",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC ID object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LCID"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/issueBL": {
      "post": {
        "summary": "Issue BL",
        "description": "Issue BL",
        "operationId": "issueBL",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LCID"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },

    "/transferBL": {
      "post": {
        "summary": "Transfer BL",
        "description": "Transfer BL",
        "operationId": "transferBL",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LCID"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },

    "/getLC": {
      "post": {
        "summary": "Get LC",
        "description": "Get LC",
        "operationId": "getLC",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LCID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/LC"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    },
    "/getLCHistory": {
      "post": {
        "summary": "Get LC History",
        "description": "Get LC History",
        "operationId": "getLCHistory",
        "produces": ["application/json"],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "LC object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LCID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/LC"
              }
            }
          },
          "400": {
            "description": "Invalid status value"
          }
        }
      }
    }
  },
  "definitions": {
    "LC": {
      "type": "object",
      "properties": {
        "lcId": {
          "type": "string",
          "description": "LC ID"
        },
        "BLId": {
          "type": "string",
          "description": "BL ID"
        },
        "pId": {
          "type": "string",
          "description": "Purchase ID"
        },
        "blOwner": {
          "type": "string",
          "description": "BL Owner"
        },
        "expiryDate": {
          "type": "string",
          "description": "Expiry Date"
        },
        "buyer": {
          "type": "string",
          "description": "Buyer Id"
        },
        "buyerBank": {
          "type": "string",
          "description": "Buyer Bank Id"
        },
        "sellerBank": {
          "type": "string",
          "description": "Seller Bank Id"
        },
        "seller": {
          "type": "string",
          "description": "Seller ID"
        },
        "amount": {
          "type": "number",
          "description": "Amount"
        },
        "status": {
          "type": "string",
          "description": "status"
        }
      }
    },
    "LCID": {
      "type": "object",
      "properties": {
        "lcId": {
          "type": "string",
          "description": "LC ID"
        }
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}